inital commit :dog:
Using the google sheets API, particiapte enables all students to answer in-class questions anonymously. The program automatically assigns participation points to students based on their answers. It also has a randomizer and picks a random student (I haven't used this feature in sections becuase I think it might create unnecessary anxiety ... but it's nice to know that I always have the option to do so! :angel: